Adds file deletion confirmation

+++ SolutionsUserOptionsCleaner/Program.cs
@@ -1,4 +1,6 @@
 ﻿using System;
+using System.Collections.Generic;
+using System.ComponentModel;
 using System.IO;
 using System.Linq;
 
@@ -14,7 +16,47 @@ namespace SolutionsUserOptionsCleaner
                 return;
 
             FindSuoFiles(args);
-            DeleteFiles();
+            DisplayFoundFiles();
+            if (FilesWereFound() && DeleteConfirmed())
+                DeleteFiles();
+            else
+                Console.ReadKey();
+        }
+
+        private static bool FilesWereFound()
+        {
+            return _filesToDelete.Length> 0;
+        }
+
+        private static bool DeleteConfirmed()
+        {
+            bool? confirmed = null;
+            while (confirmed == null)
+            {
+                Console.WriteLine("Delete files? (Y/n)");
+                var key = Console.ReadKey(true);
+                confirmed = Confirm(key);
+            }
+
+            return confirmed.Value;
+        }
+
+        private static bool? Confirm(ConsoleKeyInfo key)
+        {
+            var authorise = new List<ConsoleKey> {ConsoleKey.Enter, ConsoleKey.Y};
+            var decline = new List<ConsoleKey> {ConsoleKey.N};
+            if (authorise.Contains(key.Key))
+                return true;
+            if (decline.Contains(key.Key))
+                return false;
+            return null;
+        }
+
+        private static void DisplayFoundFiles()
+        {
+            Console.WriteLine($"Found {_filesToDelete.Length} files to delete");
+            foreach (var file in _filesToDelete)
+                Console.WriteLine($"-\t{file}");
         }
 
         private static void DeleteFiles()
